[runtime] backend and executor checker
This adds two method if backend is "cpu" and executor is "Linear".

This is required to test dynamic tensors, which is being developed for cpu and linear only. If the test is run for, e.g., "acl_neon", or "DataFlow", test will be skipped.

ONE-DCO-1.0-Signed-off-by: hyunsik-yoon <hyunsik.yoon.1024@gmail.com>