[onert] Set default backend to 'cpu' (#1662)
The first backend in BACKENDS is the backend that has the highest
priority. And 'acl_cl' has been the first for the convenience of
testing at the early stages of this project. But we should set it
to 'cpu' as it is the most general(compatible) backend.

ONE-DCO-1.0-Signed-off-by: Hanjoung Lee <hanjoung.lee@samsung.com>