[onert] Fix circular dependency of ExecutorMap
Having `shared_ptr` in `controlflow::KernelGenerator` causes circular
dependency. So this commit changes it to be a raw pointer. We may make
it to use `weak_ptr` or make it as a `unique_ptr` later.

Circular dependency was:

```
ExecutorMap -> ... -> LoweredGraph -> BackendContexts -> ... ->
controlflow::KernelGenerator -> ExecutorMap
```

For #2537

ONE-DCO-1.0-Signed-off-by: Hanjoung Lee <hanjoung.lee@samsung.com>