[onert] Introduce controlflow::UserTensor
Introduce `backend::controlflow::UserTensor` for model input and output
tensors. All the input/output tensors are going to be wrapped with
`UserTensor` so it can be treated the same way with others.

It is to remove explicit copies(Source/Sink) when starting execution.

Draft: #1325

ONE-DCO-1.0-Signed-off-by: Hanjoung Lee <hanjoung.lee@samsung.com>