[onert] Throw when tensor_registry is nullptr
This changes assert to throw when tensor_register == nullptr.

When we run some model with several subgraphs with cpu backend, if BACKEND=cpu is not given by mistake, assert checks nullptr. However, with release build, onert keeps running and dies with segmentation falut. So changing assert to throw seems better to prevent this mistake.

Signed-off-by: Hyun Sik Yoon <hyunsik.yoon.1024@gmail.com>