[onert] Optimize DynamicTensorManager::deallocInput()
This optimizes DynamicTensorManager::deallocInput().

Background -----

When handling a model with dynamic tensors, deallocInput() is always called
after running a kernel of op.

If a kernel is small, overhead of [ deallocInput() * number of ops ] could
be relatively big.

For this reason, optimization of deallocInput() is necessary.

How to -----

1. Eliminate usage of tensor index
2. This also removes the calling number of getITensor(), resulting in more optimization.

Result ----
(assuming b05eeec7 performance result is 100)

   case         |  b05eeec7  | with #4277 | more with this commit
----------------|------------|------------|----------------------
getITensors()   |     100    |     45.4   |   32.3
deallocInputs() |     100    |     68.9   |   58.8

Signed-off-by: Hyun Sik Yoon <hyunsik.yoon.1024@gmail.com>