[onert] Fix when an operand that is input & output
Insert a Permute operation and an operand to make an input can be copied
to an output. This is necessary since our API lets users give different
buffer pointers even though a tensor is an input and an output at the
same time.

- Introduce OddOutputPass to handle when an output is also an input
- Fix controlflow::DynamicTensorManager bug
- Add restriction to `PermutationInsertionPass` so the inserted Permute
  op can never be gone
- Add two tests

ONE-DCO-1.0-Signed-off-by: Hanjoung Lee <hanjoung.lee@samsung.com>