[onert] Change 'backend' field of chrome tracing
This changes `backend` field of chrome tracing to `category` and
put subgraph and backend info into `category`.

Consider we have a model that has 2 subgraphs that runs on CPU.
Only with `backend`, CPU backend of subgraph A and suggraph B cannot
be distinguished.
So now it has subgraph info as well as backend info.

For subgraph level, previously `backend` contains `"Runtime"` string
which is not quite right. So I changed the name to `category`.

Signed-off-by: Hyun Sik Yoon <hyunsik.yoon.1024@gmail.com>