[onert] Make ctor of ir::Graph take subgraph index param
Make ctor of ir::Graph take subgraph index param.

Background:
To produce right profiling information about multiple subgraphs
at execution time, each executor needs to know which 'subgraph'
they are executing. The proper place to keep the subgraph index
seems to be ir::Graph.
The subgraph index will be passed to EventObservers and displayed
to users.
If we don't want to keep the index in Graph, we need a map of
subgraph index and Graph object, which is created at loading time
and passed all the way to execution time, which seems complicated.

Signed-off-by: Hyun Sik Yoon <hyunsik.yoon.1024@gmail.com>