[luci] Introduce CopyLocoItemsToCirclePass in luci
Until refactoring for removing dependency between `loco` shape and circle shape is finished,
shape and dtype of `loco` should be copied to circle nodes to keep consistency.
This commit will introduce `CopyLocoItemsToCirclePass` and this pass will do it.
FYI, be careful that when copying `loco` shape to circle node, shape signature of the circle node
is considered, but shape of `loco` is not changed.

When the refactoring is finished, this pass will be removed.

ONE-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>