[luci] Change candidates for circle type inference pass
Until now, `loco::postorder_traversal` and `loco::output_nodes` were used for getting candidates.
However, some nodes were not inferred even they are not dead nodes.
For example, `While_002` test has following graph.
```
              +-- CircleWhileOut(ofm1) -- CircleOutput
CircleWhile --+
              +-- CircleWhileOut(ofm2)
```
In the above graph, `ofm2` is not used for graph output but it is not dead node because it is essential virtual circle node.
As a result, it is not checked by current traversal and then it becomes to have invalid type.
To prevent this, this commit will change the traversal method.

ONE-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>