[luci] Remove redundant constant creation in ResolveCustomOpMatMulPass
If there is no `bias` in `FullyConnected`, there are two ways to represent it.

- Create constant node and fill it with 0
- Create `CircleOutputExclude` and do nothing

Until now, first way is used but it is inefficient and unknown dimension should be considered.
Therefore this commit will change it to second way, which is more efficient and simple.

ONE-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>