[luci] Introduce substitutive function of loco shape/dtype getter
To deprecate `TypeInferencePass` and `ShapeInferencePass`, following functions should be removed.

- shape_get
- shape_known
- dtype_get
- dtype_known

Until now, replacing them as `rank` and `dim` of `CircleNode` was used.
However, `CircleShapeInferenceRule.cpp` and `CircleTypeInferenceRule.cpp` are using them so much.
As a result, if we do not introduce substitutive functions for them, code diff will be a lot.
Therefore, this commit will introduce substitutive functions for shape and dtype getter of loco.

ONE-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>