[onert/pareto_scheduler_app]: Add pareto scheduling application for onert
Add pareto scheduling application for onert, targeting multi-model execution on an end device. The application allows a onert session to dynamically reconfigure backends along an optimal tradeoff  in response to degradation in performance. Additionally, it allows a session to revert back to low latency backend allocation when a competing session finishes execution and frees up resources.

The application source and header files can be found under `runtime/onert/sample/pareto_scheduler`.

Reconfiguring backends on-the-fly requires modification to nnfw-specific source files and headers, all of which are included in this commit.

Signed-off-by: venkat.iyer venkat.iyer@samsung.com
Signed-off-by: venkat.iyer <venkat.iyer@samsung.com>