[onert] Remove _has_info variable from TrainingInfo
This commit removes the has_info variable from TrainingInfo.
The default value of trainnig_info in CompilerFactory.create() is nullptr.
Therefore, if create() is not called with TrainingInfo, it is the same with shouldTrain() == false.
Conversely, if create() is called with TrainingInfo, it is the same with shouldTrain() == true.

ONE-DCO-1.0-Signed-off-by: Jiyoung Yun <jy910.yun@samsung.com>