[tflchef] Refactor check_custom_op_value in OpUtils
This commit is for refactoring check_custom_op_value function in OpUtils.
First, the extype is used for a while and will be removed later.
Second, it is unnecessary to check the state of custom_code because the default value always exists.

ONE-DCO-1.0-Signed-off-by: SeungHui Lee <shsh1004.lee@samsung.com>