[onert] Rename Loss operation's printed name.
Loss is a kind of operation. It has name() member funciton which is used
to print in log. It has space (e.g. "MeanSquaredError Loss", ...).
When it is printed in IR, it looks like two words: e.g.

[   GraphDumper  ] {
[   GraphDumper  ]   %4 = @0_FullyConnected(%0, %1, %?)
[   GraphDumper  ]   %5 = @1_FullyConnected(%4, %2, %?)
[   GraphDumper  ]   %6 = @2_FullyConnected(%5, %3, %?)
[   GraphDumper  ]   %8 = @3_MeanSquaredError Loss(%6, %7)
[   GraphDumper  ] {

It would be good to remove space in operation name.

ONE-DCO-1.0-Signed-off-by: Sanggyu Lee <sg5.lee@samsung.com>