[onert] Apply reduction type to Loss
This commit applies a reduction type to the Loss operation.
It changes the loss output shape to match the batch size using a static
shape inferer. The loss information is stored in the TrainingExecutor and
the loss value is returned according to the reduction type.

ONE-DCO-1.0-Signed-off-by: Jiyoung Yun <jy910.yun@samsung.com>