[onert][tools/onert_train] Use num_of_trainable_ops
This commit change way of indicating trainable operations. In the previous version
it was an array with indexes of operations.
Currently it is a number of operations to be trained from the back.
The change has an infuence on onert_train and test model generation.

ONE-DCO-1.0-Signed-off-by: Mateusz Bencer <m.bencer@partner.samsung.com>