[compute/cker] Introduce compile-time definitions file
This commit introduces a modern C++ version of compile-time features detection utilities. The advantage of this approach over using #ifdef is mostly safety which C++ itself provides. The main idea for the usage of those checks is with `if constexpr` and in the context of SFINAE.

ONE-DCO-1.0-Signed-off-by: Tomasz Dolbniak t.dolbniak@partner.samsung.com