[infra/debian] Revise rules for circle-interpreter
This removes the ONE build scripts, as the `circle-interpreter` and
its related binaries will be built in the GitHub workflow.
This is necessary because Debian builds on Launchpad do not support
fetching external code, such as TensorFlow, during the build process.

ONE-DCO-1.0-Signed-off-by: Seungho Henry Park <shs.park@samsung.com>