[onert] Refactor tensorinfo to be more pythonic and robust
This commit refactors the tensorinfo structure to be easier to use on
the Python side. Instead of exporting C dimension array which needs to
be trimmed on the Python side, now it exports Python tuple with length
equal to tensor rank. Also, the entire structure is designed to be
immutable to improve the understanding of the API concept - altering
the info structure will not change the tensor.

The new tensorinfo structure consists of a data type and the tensor
shape. The data type is a dedicated type exposed by the native Python
library. It provides a direct mapping between numpy data type and the
internal NNFW data type. Additionally, it allows to expose all internal
data types (quantized types) which was not possible with previous
approach - there is no 1:1 mapping between C++ types and NNFW types.

ONE-DCO-1.0-Signed-off-by: Arkadiusz Bokowy <a.bokowy@samsung.com>