[circle-mlir] Discard output from the cd command when followed by pwd
This commits fixes the issue when `cd` is followed by the `pwd` in a
bash one-liner which should get the directory in which the executed
script is located. The problem with such approach is that the `cd`
command can print the destination directory in case when one has the
CDPATH environmental variable set.

ONE-DCO-1.0-Signed-off-by: Arkadiusz Bokowy <a.bokowy@samsung.com>