[onert] Rename the python API directory
This PR renames the directory in which the python API files are located. It adjusts the setup.py file accordingly to maintain the ability to build the wheel using this method.

This change is related to the draft https://github.com/Samsung/ONE/pull/16192 in which this directory has been renamed so that it's easier to pack the python files using the new method but without the need of copying those files to a temporary directory.

Separating the renaming from the other changes will make the final PR introducing pyproject.toml much smaller and easier to review.

ONE-DCO-1.0-Signed-off-by: Tomasz Dołbniak t.dolbniak@partner.samsung.com